+++ libs/conversation-view/src/components/AdvancedView/TableSettings/AgGridColumnPanel/useAgGridColumnPreferences.ts
@@ -3,7 +3,7 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import type { GridApi } from 'ag-grid-community';
 import type { AgGridInitialColumnsState } from './types';
-import { captureInitialColumnsState } from './helpers';
+import { captureInitialColumnsState } from './helpers/columnStateSnapshot';
 import { useAgGridColumnGridListeners } from './useAgGridColumnGridListeners';
 
 export function useAgGridColumnPreferences({
